AlexaバックエンドをローカルでDocker管理できるように修正
